Apply suggestions from code review

+++ numpy/core/src/multiarray/datetime_busdaycal.c
@@ -73,7 +73,7 @@ PyArray_WeekMaskConverter(PyObject *weekmask_in, npy_bool *weekmask)
 general_weekmask_string:
         /* a string like "SatSun" or "Mon Tue Wed" */
         memset(weekmask, 0, 7);
-        for (int i = 0; i < len; i += 3) {
+        for (Py_ssize_t i = 0; i < len; i += 3) {
             while (isspace(str[i]))
                 ++i;
